BMC#11344 Template keyword highlight
Added highlighting of certain keywords to descriptions. The keywords are taken
from http://wiki.meego.com/TestCaseTemplate and are hard-coded in the template
as XSLT has quite limited support for e.g. lists. In addition XSLT 1.0 does
not have regular expressions support so the keywords are detected by string
matching and thus there's no room for deviations in the template.