BMC#11478 List planned cases separately
Modified the XSLT to report cases with no state or state other then Design|design in the "Test cases" section, and cases with state Design|design separately in "Planned cases" section. The output of these two sections is the same.

Currently this thus lists suite and set information twice if a suite/set has both planned and other cases. This can of course be changed if needed, will consult the feature request initiator.