Google Chrome fix
For some odd reasong Chrome breaks when it finds < from the script section,
thus changed to while loop. For documentations sake: outputting the sign
with xsl:value-of it can be laid out in a way that works for Chrome but
then IE breaks (it does not understand &lt; in the output). This is likely
some iffy in Chrome since if I ran the transformation with Sablotron,
Chrome naturally did not break in the < in JavaScript.